backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
ultralytics
Pillow
